+++ src/os2datascanner/projects/report/reportapp/admin.py
@@ -9,6 +9,7 @@ from django.contrib.auth.models import User
 from django.template.response import TemplateResponse
 
 from .models.documentreport import DocumentReport
+from .models.scanner_reference import ScannerReference
 
 from os2datascanner.engine2.pipeline import messages  # noqa
 from os2datascanner.projects.report.organizations.models import Organization, Account
@@ -16,26 +17,6 @@ from os2datascanner.projects.report.reportapp.management.commands import (
         result_collector as r_c)
 
 import os2datascanner.engine2.model._staging.sbsysdb  # noqa
-
-
-class ScannerJobFilter(admin.SimpleListFilter):
-    title = _("scanner job")
-
-    parameter_name = "scanner"
-
-    def lookups(self, request, model_admin):
-        # Get all (scanner job name, scanner primary key) tuples
-        objects = DocumentReport.objects.order_by("-scanner_job_pk").only(
-                "scanner_job_name", "scanner_job_pk").distinct(
-                "scanner_job_name", "scanner_job_pk")
-        return (
-            (o.scanner_job_pk, o.scanner_job_name) for o in objects
-        )
-
-    def queryset(self, request, queryset):
-        v = self.value()
-        if v is not None:
-            return queryset.filter(scanner_job_pk=v)
 
 
 @admin.register(DocumentReport)
@@ -95,14 +76,14 @@ class DocumentReportAdmin(admin.ModelAdmin):
         'name',
         'number_of_matches',
         'organization',
-        'scanner_job_name',
+        'scanner_job',
         'aliases',
         'resolution_status',
         'resolution_time', 'only_notify_superadmin')
 
     list_filter = (
         'organization',
-        ScannerJobFilter,)
+        'scanner_job',)
 
     def aliases(self, dr):
         return ", ".join([alias_relation.account.username
@@ -111,6 +92,19 @@ class DocumentReportAdmin(admin.ModelAdmin):
     def get_queryset(self, request):
         return super(DocumentReportAdmin, self).get_queryset(request).select_related(
             'organization').prefetch_related('alias_relation')
+
+
+@admin.register(ScannerReference)
+class ScannerReferenceAdmin(admin.ModelAdmin):
+    list_display = (
+        "scanner_name",
+        "scanner_pk",
+        "organization"
+    )
+
+    list_filter = (
+        "organization",
+    )
 
 
 class ProfileInline(admin.TabularInline):