+++ src/os2datascanner/projects/report/reportapp/management/commands/event_collector.py
@@ -27,6 +27,7 @@ from os2datascanner.engine2.pipeline.utilities.pika import PikaPipelineThread
 from os2datascanner.projects.report.organizations.models import (Account, Alias, Organization,
                                                                  OrganizationalUnit, Position)
 from os2datascanner.projects.report.reportapp.models.documentreport import DocumentReport
+from os2datascanner.projects.report.reportapp.models.scanner_reference import ScannerReference
 from prometheus_client import Summary, start_http_server
 from ...utils import create_alias_and_match_relations
 
@@ -63,6 +64,25 @@ def event_message_received_raw(body):  # noqa: CCR001 C901
 
                     else:
                         logger.info("No objects to create of", model=model.__name__)
+
+                if raw_model_data := classes.get("Scanner"):
+                    # We have received a creation event of Scanner objects from the Admin module.
+                    # The Report module doesn't have a Scanner model,
+                    # but we instead need to create ScannerReferences.
+                    for scanner_dict in raw_model_data:
+                        sr = ScannerReference.objects.create(
+                            scanner_pk=scanner_dict['pk'],
+                            scanner_name=scanner_dict['name'],
+                            only_notify_superadmin=scanner_dict['only_notify_superadmin'],
+                            scan_entire_org=scanner_dict['scan_entire_org'],
+                        )
+                        if org_uuid := scanner_dict.get('organization'):
+                            sr.organization = Organization.objects.get(uuid=org_uuid)
+                        ou_uuids = scanner_dict.get('org_unit', [])
+                        sr.org_units.set(OrganizationalUnit.objects.filter(uuid__in=ou_uuids))
+                        sr.save()
+
+                    logger.info("Successfully ran broadcast create!")
 
             elif event_type == "bulk_event_update":
                 logger.info("Initiating broadcast update transaction...")
@@ -103,6 +123,28 @@ def event_message_received_raw(body):  # noqa: CCR001 C901
                         logger.info("Successfully ran broadcast update!")
                     else:
                         logger.info("Nothing to update for", model=model.__name__)
+
+                if raw_model_data := classes.get("Scanner"):
+                    # We have received an update event of Scanner objects from the Admin module.
+                    # The Report module doesn't have a Scanner model,
+                    # but we instead need to update ScannerReferences.
+                    for scanner_dict in raw_model_data:
+                        sr, created = ScannerReference.objects.get_or_create(
+                            scanner_pk=scanner_dict['pk'])
+                        if created:
+                            logger.info("Update for unknown ScannerReference received.")
+
+                        sr.scanner_name = scanner_dict['name']
+                        sr.only_notify_superadmin = scanner_dict['only_notify_superadmin']
+                        sr.scan_entire_org = scanner_dict['scan_entire_org']
+
+                        if org_uuid := scanner_dict.get('organization'):
+                            sr.organization = Organization.objects.get(uuid=org_uuid)
+                        ou_uuids = scanner_dict.get('org_unit', [])
+                        sr.org_units.set(OrganizationalUnit.objects.filter(uuid__in=ou_uuids))
+                        sr.save()
+
+                    logger.info("Successfully ran broadcast update!")
 
             elif event_type == "bulk_event_delete":
                 for model in ORDER_OF_DELETION:
@@ -156,14 +198,14 @@ def handle_clean_account_message(body):
         account_usernames = account_dict.get("usernames")
 
         related_reports = DocumentReport.objects.filter(
-            alias_relation__account__in=account_uuids, scanner_job_pk=scanner_pk)
+            alias_relation__account__in=account_uuids, scanner_job__scanner_pk=scanner_pk)
 
         _, deleted_reports_dict = related_reports.delete()
         deleted_reports = deleted_reports_dict.get("os2datascanner_report.DocumentReport", 0)
 
         logger.info(
             "Deleted DocumentReport objects!",
-            count=deleted_reports, scanner_job_pk=scanner_pk,
+            count=deleted_reports, scanner_pk=scanner_pk,
             associated_accounts=', '.join(account_usernames))
 
 
@@ -173,17 +215,17 @@ def handle_clean_problem_message(body):
     logger.info("CleanProblemMessage published by",
                 publisher=body.get('publisher'), published_time=body.get('time'))
 
-    scanners = body.get("scanners", [])
+    scanners = ScannerReference.objects.filter(scanner_pk__in=body.get("scanners", []))
 
     related_problems = DocumentReport.objects.filter(
-        number_of_matches=0, scanner_job_pk__in=scanners)
+        number_of_matches=0, scanner_job__in=scanners)
     _, deleted_problems_dict = related_problems.delete()
     deleted_problems = deleted_problems_dict.get(
         "os2datascanner_report.DocumentReport", 0)
 
     logger.info(
         "Deleted DocumentReport objects without matches ",
-        count=deleted_problems, scanner_job_pk=scanners)
+        count=deleted_problems, scanner_job=scanners)
 
 
 class EventCollectorRunner(PikaPipelineThread):